Post - Game Changes
1) Sov Buffs
- +25% Retention in the Wartime Propaganda Focus
(Sov currently can't switch to newer eq in 42, this should rectify the issue)
- Sov late game operations have air bonuses 1943+
(Speer on Germany is generally speaking too strong. Instead of nerfing, slowly tuning Sov air to be on par was chosen)
- Sov Worker-Peasants Red Army eq bonus changed from -15 to -22,5%
(Currently, against proper axis Soviet's can't recover the losses fast enough. When Soviet tanks get stats, the infantry is so out of equipment that new mills can't be assigned on tanks)
2) Checksum Change